+++ source/patterns/@aws-solutions-constructs/core/test/sns-helper.test.ts
@@ -18,21 +18,44 @@ import { SynthUtils, expect as expectCDK, haveResource } from '@aws-cdk/assert';
 import '@aws-cdk/assert/jest';
 
 // --------------------------------------------------------------
-// Test deployment with no properties
+// Test deployment with no properties using AWS Managed KMS Key
 // --------------------------------------------------------------
-test('Test deployment with no properties', () => {
+test('Test deployment with no properties using AWS Managed KMS Key', () => {
     // Stack
     const stack = new Stack();
     // Helper declaration
-    defaults.buildTopic(stack);
+    defaults.buildTopic(stack, {});
     // Assertion 1
     expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
+    // Assertion 2
+    expect(stack).toHaveResource("AWS::SNS::Topic", {
+      KmsMasterKeyId: {
+        "Fn::Join": [
+          "",
+          [
+            "arn:",
+            {
+              Ref: "AWS::Partition"
+            },
+            ":kms:",
+            {
+              Ref: "AWS::Region"
+            },
+            ":",
+            {
+              Ref: "AWS::AccountId"
+            },
+            ":alias/aws/sns"
+          ]
+        ]
+      }
+    });
 });
 
 // --------------------------------------------------------------
-// Test deployment w/ custom properties
+// Test deployment without imported encryption key
 // --------------------------------------------------------------
-test('Test deployment w/ custom properties', () => {
+test('Test deployment without imported encryption key', () => {
   // Stack
   const stack = new Stack();
   // Helper declaration
@@ -40,29 +63,7 @@ test('Test deployment w/ custom properties', () => {
       topicProps: {
           topicName: "custom-topic"
       },
-      enableEncryption: true
-  });
-  // Assertion 1
-  expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
-  // Assertion 2
-  expect(stack).toHaveResource("AWS::SNS::Topic", {
-      TopicName: "custom-topic"
-  });
-});
-
-// --------------------------------------------------------------
-// Test deployment w/ imported encryption key
-// --------------------------------------------------------------
-test('Test deployment w/ imported encryption key', () => {
-  // Stack
-  const stack = new Stack();
-  // Helper declaration
-  defaults.buildTopic(stack, {
-      topicProps: {
-          topicName: "custom-topic"
-      },
-      enableEncryption: true,
-      encryptionKey: defaults.buildEncryptionKey(stack)
+      enableEncryptionWithCustomerManagedKey: true
   });
   // Assertion 1
   expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
@@ -72,86 +73,116 @@ test('Test deployment w/ imported encryption key', () => {
   });
   // Assertion 3
   expect(stack).toHaveResource("AWS::KMS::Key", {
-      EnableKeyRotation: true
+    EnableKeyRotation: true
+  });
+});
+
+// --------------------------------------------------------------
+// Test deployment w/ imported encryption key
+// --------------------------------------------------------------
+test('Test deployment w/ imported encryption key', () => {
+  // Stack
+  const stack = new Stack();
+  // Generate KMS Key
+  const key = defaults.buildEncryptionKey(stack);
+  // Helper declaration
+  defaults.buildTopic(stack, {
+      topicProps: {
+          topicName: "custom-topic"
+      },
+      enableEncryptionWithCustomerManagedKey: true,
+      encryptionKey: key
+  });
+  // Assertion 1
+  expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
+  // Assertion 2
+  expect(stack).toHaveResource("AWS::SNS::Topic", {
+    KmsMasterKeyId: {
+      "Fn::GetAtt": [
+        "EncryptionKey1B843E66",
+        "Arn"
+      ]
+    },
+    TopicName: "custom-topic"
   });
 });
 
 test('Check SNS Topic policy', () => {
     const stack = new Stack();
-    defaults.buildTopic(stack);
+    defaults.buildTopic(stack, {});
 
     expectCDK(stack).to(haveResource("AWS::SNS::TopicPolicy", {
-        PolicyDocument: {
-            Statement: [
-              {
-                Action: [
-                  "SNS:Publish",
-                  "SNS:RemovePermission",
-                  "SNS:SetTopicAttributes",
-                  "SNS:DeleteTopic",
-                  "SNS:ListSubscriptionsByTopic",
-                  "SNS:GetTopicAttributes",
-                  "SNS:Receive",
-                  "SNS:AddPermission",
-                  "SNS:Subscribe"
-                ],
-                Condition: {
-                  StringEquals: {
-                    "AWS:SourceOwner": {
-                      Ref: "AWS::AccountId"
-                    }
-                  }
-                },
-                Effect: "Allow",
-                Principal: {
-                  AWS: {
-                    "Fn::Join": [
-                      "",
-                      [
-                        "arn:",
-                        {
-                          Ref: "AWS::Partition"
-                        },
-                        ":iam::",
-                        {
-                          Ref: "AWS::AccountId"
-                        },
-                        ":root"
-                      ]
-                    ]
-                  }
-                },
-                Resource: {
-                  Ref: "SnsTopic2C1570A4"
-                },
-                Sid: "TopicOwnerOnlyAccess"
-              },
-              {
-                Action: [
-                  "SNS:Publish",
-                  "SNS:RemovePermission",
-                  "SNS:SetTopicAttributes",
-                  "SNS:DeleteTopic",
-                  "SNS:ListSubscriptionsByTopic",
-                  "SNS:GetTopicAttributes",
-                  "SNS:Receive",
-                  "SNS:AddPermission",
-                  "SNS:Subscribe"
-                ],
-                Condition: {
-                  Bool: {
-                    "aws:SecureTransport": "false"
-                  }
-                },
-                Effect: "Deny",
-                Principal: "*",
-                Resource: {
-                  Ref: "SnsTopic2C1570A4"
-                },
-                Sid: "HttpsOnly"
-              }
+      PolicyDocument: {
+        Statement: [
+          {
+            Action: [
+              "SNS:Publish",
+              "SNS:RemovePermission",
+              "SNS:SetTopicAttributes",
+              "SNS:DeleteTopic",
+              "SNS:ListSubscriptionsByTopic",
+              "SNS:GetTopicAttributes",
+              "SNS:Receive",
+              "SNS:AddPermission",
+              "SNS:Subscribe"
             ],
-            Version: "2012-10-17"
-        },
+            Condition: {
+              StringEquals: {
+                "AWS:SourceOwner": {
+                  Ref: "AWS::AccountId"
+                }
+              }
+            },
+            Effect: "Allow",
+            Principal: {
+              AWS: {
+                "Fn::Join": [
+                  "",
+                  [
+                    "arn:",
+                    {
+                      Ref: "AWS::Partition"
+                    },
+                    ":iam::",
+                    {
+                      Ref: "AWS::AccountId"
+                    },
+                    ":root"
+                  ]
+                ]
+              }
+            },
+            Resource: {
+              Ref: "SnsTopic2C1570A4"
+            },
+            Sid: "TopicOwnerOnlyAccess"
+          },
+          {
+            Action: [
+              "SNS:Publish",
+              "SNS:RemovePermission",
+              "SNS:SetTopicAttributes",
+              "SNS:DeleteTopic",
+              "SNS:ListSubscriptionsByTopic",
+              "SNS:GetTopicAttributes",
+              "SNS:Receive",
+              "SNS:AddPermission",
+              "SNS:Subscribe"
+            ],
+            Condition: {
+              Bool: {
+                "aws:SecureTransport": "false"
+              }
+            },
+            Effect: "Deny",
+            Principal: "*",
+            Resource: {
+              Ref: "SnsTopic2C1570A4"
+            },
+            Sid: "HttpsOnly"
+          }
+        ],
+        Version: "2012-10-17"
+      },
     }));
   });