+++ source/patterns/@aws-solutions-constructs/core/test/sqs-helper.test.ts
@@ -25,7 +25,7 @@ test('Test minimal deployment with no properties', () => {
     // Stack
     const stack = new Stack();
     // Helper declaration
-    defaults.buildQueue(stack, 'primary-queue');
+    defaults.buildQueue(stack, 'primary-queue', {});
     // Assertion 1
     expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
 });
@@ -58,7 +58,7 @@ test('Test dead letter queue deployment/configuration', () => {
     const stack = new Stack();
     // Helper setup
     const encKey = defaults.buildEncryptionKey(stack);
-    const dlq = defaults.buildQueue(stack, 'dead-letter-queue');
+    const [dlq] = defaults.buildQueue(stack, 'dead-letter-queue', {});
     const dlqi = defaults.buildDeadLetterQueue({
         deadLetterQueue: dlq,
         maxReceiveCount: 3
@@ -84,7 +84,7 @@ test('Test dead letter queue deployment/configuration w/o mrc', () => {
   const stack = new Stack();
   // Helper setup
   const encKey = defaults.buildEncryptionKey(stack);
-  const dlq = defaults.buildQueue(stack, 'dead-letter-queue');
+  const [dlq] = defaults.buildQueue(stack, 'dead-letter-queue', {});
   const dlqi = defaults.buildDeadLetterQueue({
       deadLetterQueue: dlq
   });
@@ -108,7 +108,7 @@ test('Test existingQueueObj', () => {
   // Stack
   const stack = new Stack();
   // Helper setup
-  const existingQueue = defaults.buildQueue(stack, 'existing-queue', {
+  const [existingQueue] = defaults.buildQueue(stack, 'existing-queue', {
     queueProps: {
       queueName: 'existing-queue'
     }
@@ -119,4 +119,51 @@ test('Test existingQueueObj', () => {
   });
   // Assertion 1
   expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
+});
+
+// --------------------------------------------------------------
+// Test deployment w/ imported encryption key
+// --------------------------------------------------------------
+test('Test deployment w/ imported encryption key', () => {
+  // Stack
+  const stack = new Stack();
+  // Helper declaration
+  defaults.buildQueue(stack, 'existing-queue', {
+    queueProps: {
+      queueName: 'existing-queue'
+    },
+    enableEncryptionWithCustomerManagedKey: true,
+    encryptionKey: defaults.buildEncryptionKey(stack)
+  });
+  // Assertion 1
+  expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
+  // Assertion 2
+  expect(stack).toHaveResource("AWS::SQS::Queue", {
+    QueueName: "existing-queue"
+  });
+  // Assertion 3
+  expect(stack).toHaveResource("AWS::KMS::Key", {
+    EnableKeyRotation: true
+  });
+});
+
+// --------------------------------------------------------------
+// Test deployment without imported encryption key
+// --------------------------------------------------------------
+test('Test deployment without imported encryption key', () => {
+  // Stack
+  const stack = new Stack();
+  // Helper declaration
+  defaults.buildQueue(stack, 'existing-queue', {
+    queueProps: {
+      queueName: 'existing-queue'
+    }
+  });
+  // Assertion 1
+  expect(SynthUtils.toCloudFormation(stack)).toMatchSnapshot();
+  // Assertion 2
+  expect(stack).toHaveResource("AWS::SQS::Queue", {
+    QueueName: "existing-queue",
+    KmsMasterKeyId: "alias/aws/sqs"
+  });
 });